+++ samples/src/main/java/in/testpress/samples/MainActivity.java
@@ -14,7 +14,7 @@ import in.testpress.samples.course.CourseSampleActivity;
 import in.testpress.samples.store.StoreSampleActivity;
 
 import static in.testpress.samples.core.TestpressCoreSampleActivity.AUTHENTICATE_REQUEST_CODE;
-import static in.testpress.util.PermissionUtilKt.mayBeAskAllPermissions;
+import static in.testpress.util.extension.ActivityKt.askAllPermissions;
 
 public class MainActivity extends BaseToolBarActivity {
 
@@ -22,7 +22,7 @@ public class MainActivity extends BaseToolBarActivity {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        mayBeAskAllPermissions(this);
+        askAllPermissions(this);
         findViewById(R.id.core).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
